fix(type-helpers): make type-test assertions gate the build
union-to-intersection.d.test.ts was parsed by TypeScript as an ambient
declaration file, rendering all @ts-expect-error / @ts-expect-no-error
assertions inert during `heft build`. Rename to .test.ts so the file
is a real module whose assertions are type-checked.

Add .npmignore to keep the test source and compiled artifacts out of
the published tarball (@rhombus-toolkit/type-helpers shouldPublish:
true in rush.json).

Same fix as #269 (typed-pluralizer).
